mgr/dashboard: Relocate empty pipe
Move empty pipe into shared/pipes directory.
