+++ src/pybind/mgr/dashboard/frontend/src/app/shared/pipes/empty.pipe.ts
